python/requirements.txt
yfinance
python-dotenv
python-binance
requests
pandas